We use BaseServer around here, we don't care about RAM costs!
